Split L.H.S.TypeCheck into multiple modules
