+++ hakaru.cabal
@@ -61,6 +61,7 @@ Library
                        Language.Hakaru.Syntax.SArgs,
                        Language.Hakaru.Syntax.Transform,
                        Language.Hakaru.Syntax.TypeCheck,
+                       Language.Hakaru.Syntax.TypeCheck.TypeCheckMonad,
                        Language.Hakaru.Syntax.TypeOf,
                        Language.Hakaru.Syntax.Uniquify,
                        Language.Hakaru.Syntax.Unroll,
@@ -185,6 +186,7 @@ Test-Suite system-testsuite
                        Language.Hakaru.Syntax.SArgs
                        Language.Hakaru.Syntax.Transform
                        Language.Hakaru.Syntax.TypeCheck
+                       Language.Hakaru.Syntax.TypeCheck.TypeCheckMonad
                        Language.Hakaru.Syntax.TypeOf
                        Language.Hakaru.Syntax.Uniquify
                        Language.Hakaru.Syntax.Unroll
@@ -285,6 +287,7 @@ Test-Suite relationships-testsuite
                        Language.Hakaru.Syntax.SArgs
                        Language.Hakaru.Syntax.Transform
                        Language.Hakaru.Syntax.TypeCheck
+                       Language.Hakaru.Syntax.TypeCheck.TypeCheckMonad
                        Language.Hakaru.Syntax.TypeOf
                        Language.Hakaru.Syntax.Uniquify
                        Language.Hakaru.Syntax.Unroll
